License Change to the GPLv3
Typically, I release my software under the Apache License. That said,
I'm making an exception here. It will be possible one day to run this on
a server without exposing the API and it will be useful for larger sites
doing that. If such a site *does* do that, and then ends up making
improvements to the parsing code or creates some new, useful
augmentation to the API, I want to ensure that's released - partly so I
don't have to recreate the work on my end.

Since this can be run as a separate application on your servers, I don't
forsee the licensing issues being a problem for any company that wishes
to provide intelligence on this data. Intelligence algorithms are things
I don't mind being private or proprietary, but basic access to data -
what this project provides - shouldn't be.